Added table for thinkof recordings to flow view, moved tableview delegate functions into parent class, inherited by view and flow view controller
